requirements.txt
flask
flask_sqlalchemy
python-dotenv
requests
flask-cors